fix logbuilder reversion rename files to remove ct_
